+++ src/components/Molecules/NavBar/NavBar.stories.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+import { MemoryRouter } from 'react-router-dom';
+import NavBar from './NavBar';
+
+storiesOf('components/Molecules/NavBar', module).add('NavBar', () => (
+  <MemoryRouter>
+    <NavBar />
+  </MemoryRouter>
+));
