+++ src/components/Molecules/TeamCarousel/TeamCarousel.stories.js
@@ -4,7 +4,5 @@ import TeamCarousel from './TeamCarousel.tsx';
 
 storiesOf('components/Molecules/TeamCarousel', module).add(
   'TeamCarousel',
-  () => {
-    return <TeamCarousel />;
-  }
+  () => <TeamCarousel />
 );
